global stats & fixes

+++ src/main/java/ServerAmministratore/GlobalStatsServices.java
@@ -40,7 +40,7 @@ public class GlobalStatsServices {
     @Produces({"application/json", "application/xml"})
     public Response getDelivMed (@PathParam("t1") int t1, @PathParam("t2") int t2) {
 
-        int med = GlobalStats.getGlobalStatsInstance().getDelivMed(t1,t2);
+        float med = GlobalStats.getGlobalStatsInstance().getDelivMed(t1,t2);
         return Response.ok(med).build();
 
     }
@@ -50,7 +50,7 @@ public class GlobalStatsServices {
     @Produces({"application/json", "application/xml"})
     public Response getKMMed (@PathParam("t1") int t1, @PathParam("t2") int t2) {
 
-        int med = GlobalStats.getGlobalStatsInstance().getKMMed(t1,t2);
+        float med = GlobalStats.getGlobalStatsInstance().getKMMed(t1,t2);
         return Response.ok(med).build();
 
     }
